# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lesson {
  id: number;
  number: number;
  title: string;
  description: string;
  totalExercises: number;
  topics: string[];
}

export type ExerciseDifficulty =
  (typeof ExerciseDifficulty)[keyof typeof ExerciseDifficulty];

export const ExerciseDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Exercise {
  id: number;
  type: string;
  title: string;
  description: string;
  lessonId: number;
  difficulty: ExerciseDifficulty;
}

export interface LessonDetail {
  id: number;
  number: number;
  title: string;
  description: string;
  theory: string;
  topics: string[];
  exercises: Exercise[];
}

export interface Pair {
  front: string;
  back: string;
  /** @nullable */
  hint?: string | null;
}

export interface ExerciseContent {
  type: string;
  title: string;
  instructions: string;
  /** @nullable */
  wordList?: string[] | null;
  /** @nullable */
  numberList?: number[] | null;
  /** @nullable */
  pairs?: Pair[] | null;
}

export interface Session {
  id: number;
  exerciseType: string;
  exerciseTitle: string;
  itemsTotal: number;
  itemsCorrect: number;
  accuracy: number;
  durationSeconds: number;
  completedAt: string;
}

export interface SessionInput {
  exerciseType: string;
  exerciseTitle: string;
  itemsTotal: number;
  itemsCorrect: number;
  durationSeconds: number;
}

export interface ExerciseBreakdown {
  exerciseType: string;
  exerciseTitle: string;
  sessionCount: number;
  averageAccuracy: number;
}

export interface DashboardStats {
  totalSessions: number;
  totalItemsPracticed: number;
  averageAccuracy: number;
  currentStreak: number;
  bestAccuracy: number;
  exerciseBreakdown: ExerciseBreakdown[];
  recentSessions: Session[];
}

export interface ProgressPoint {
  date: string;
  exerciseType: string;
  accuracy: number;
  itemsCorrect: number;
}

export interface Flashcard {
  front: string;
  back: string;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  imageCode?: string | null;
}

export interface FlashcardSet {
  type: string;
  title: string;
  description: string;
  cards: Flashcard[];
}

export type GetExercisesParams = {
  /**
   * @nullable
   */
  lessonId?: number | null;
};

export type GetSessionsParams = {
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  exerciseType?: string | null;
};

export type GetProgressStatsParams = {
  /**
   * @nullable
   */
  days?: number | null;
};