# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseBreakdown } from "./exerciseBreakdown";
import type { Session } from "./session";

export interface DashboardStats {
  totalSessions: number;
  totalItemsPracticed: number;
  averageAccuracy: number;
  currentStreak: number;
  bestAccuracy: number;
  exerciseBreakdown: ExerciseBreakdown[];
  recentSessions: Session[];
}