# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseDifficulty } from "./exerciseDifficulty";

export interface Exercise {
  id: number;
  type: string;
  title: string;
  description: string;
  lessonId: number;
  difficulty: ExerciseDifficulty;
}