# lib/api-zod/src/generated/types/exerciseBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ExerciseBreakdown {
  exerciseType: string;
  exerciseTitle: string;
  sessionCount: number;
  averageAccuracy: number;
}