# lib/api-zod/src/generated/types/exerciseContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Pair } from "./pair";

export interface ExerciseContent {
  type: string;
  title: string;
  instructions: string;
  /** @nullable */
  wordList?: string[] | null;
  /** @nullable */
  numberList?: number[] | null;
  /** @nullable */
  pairs?: Pair[] | null;
}