# lib/api-zod/src/generated/types/flashcard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Flashcard {
  front: string;
  back: string;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  imageCode?: string | null;
}