# lib/api-zod/src/generated/types/flashcardSet.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from "./flashcard";

export interface FlashcardSet {
  type: string;
  title: string;
  description: string;
  cards: Flashcard[];
}