# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: number;
  number: number;
  title: string;
  description: string;
  totalExercises: number;
  topics: string[];
}