# lib/api-zod/src/generated/types/pair.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Pair {
  front: string;
  back: string;
  /** @nullable */
  hint?: string | null;
}