# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mnemonics training platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  exerciseType: string;
  exerciseTitle: string;
  itemsTotal: number;
  itemsCorrect: number;
  durationSeconds: number;
}